l.XS

+++ MyGameProject/Game/Bat.h
@@ -13,7 +13,7 @@ public:
 	Bat();
 	~Bat();
 	void Update(float dt) override;
-	void OnCollision(Entity* impactor, Entity::SideCollision side, float collisionTime, double dt = 1.0 / 60);
+	void OnCollision(Entity* impactor, Entity::SideCollision side, float collisionTime, float dt = 1.0 / 60);
 	void SetState(EnemyState::eState state);
 	void Spawn();
 };
